+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.AI.API/Interfaces/Services/RemoteServices/IContentsApiRemoteService.cs
@@ -7,4 +7,5 @@ public interface IContentsApiRemoteService
     //Task<bool> EditContentBatchAsync(Dictionary<long, string> batchEditConentRequests, HttpRequest httpRequest);
     Task<long> AddTopic(TopicApiAddRequest payload, UserApiDto user);
     Task<bool> EditArticleAsync(ArticleApiUpsertRequest payload, UserApiDto user);
+    Task<ArticleApiDto?> GetArticleById(long articleId);
 }