+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Common/ApiDtos/ArticleApiDto.cs
@@ -1,0 +1,26 @@
+namespace FF.Articles.Backend.Common.ApiDtos;
+public class ArticleApiDto
+{
+    public long ArticleId { get; set; }
+    public DateTime? CreateTime { get; set; }
+    public DateTime? UpdateTime { get; set; }
+
+    public string Title { get; set; }
+    public string Content { get; set; }
+    public string Abstract { get; set; }
+    /// <summary>
+    /// 1: article, 2: sub article, 3: topic article
+    /// </summary>
+    public string ArticleType { get; set; }
+    public long? ParentArticleId { get; set; }
+    public List<ArticleApiDto> SubArticles { get; set; } = new();
+    public long UserId { get; set; }
+    public UserApiDto? User { get; set; }
+    public long TopicId { get; set; }
+    public string TopicTitle { get; set; } = string.Empty;
+    public List<string> Tags { get; set; } = new();
+    public int SortNumber { get; set; }
+    public int IsHidden { get; set; }
+
+}
+