+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Common/RabbitMQ/QueueList.cs
@@ -3,6 +3,6 @@ namespace FF.Articles.Backend.Common.RabbitMQ;
 public static class QueueList
 {
     public const string GenerateArticleQueue = "generate_article_queue";
-    public const string ArticleReadyQueue = "article_ready_queue";
+    public const string EditArticleQueue = "edit_article_queue";
     public const string AddArticleQueue = "add_article_queue";
 }