+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/MqConsumers/AddArticleConsumer.cs
@@ -21,7 +21,7 @@ public class AddArticleConsumer(IConnection connection, IServiceScopeFactory ser
         }
 
         using var scope = _serviceScopeFactory.CreateScope();
-        var articleService = scope.ServiceProvider.GetRequiredService<Func<string, IArticleService>>()("v1");
+        var articleService = scope.ServiceProvider.GetRequiredService<IArticleService>();
 
         var articleId = generateRequest.Id ?? 0L;
         var article = await articleService.GetByIdAsync(articleId);